requirements.txt
Django==5.2.9
djangorestframework==3.15.2
django-cors-headers==5.0.0
razorpay==1.4.1
Pillow==11.1.0
